requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-core==1.2.16
langchain-mongodb==0.11.0
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-checkpoint-mongodb==0.3.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.7
lark==1.3.1
numpy==2.4.2
openai==2.24.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
pymongo-search-utils==0.3.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
SpeechRecognition==3.14.5
SQLAlchemy==2.0.47
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
